+++ lib/leftovers/matcher_builders/node_name.rb
@@ -5,6 +5,7 @@ module Leftovers
     module NodeName
       def self.build(name_pattern)
         matcher = ::Leftovers::MatcherBuilders::Name.build(name_pattern)
+
         return unless matcher
 
         ::Leftovers::Matchers::NodeName.new(matcher)