+++ lib/leftovers/matcher_builders/string_pattern.rb
@@ -12,9 +12,6 @@ module Leftovers
           /\A#{::Regexp.escape(has_prefix)}/
         elsif has_suffix
           /#{::Regexp.escape(has_suffix)}\z/
-        # :nocov:
-        else raise
-          # :nocov:
         end
       end
     end